requirements.txt
python-binance
python-dotenv
psycopg2-binary
